Use distinct Java packages for each Maven module
- Move `modbus-tcp` classes to `com.digitalpetri.modbus.tcp`
- Move `modbus-serial` classes to `com.digitalpetri.modbus.serial`
- Move `modbus-test` classes to `com.digitalpetri.modbus.test`
- Explicitly assign Java module names rather than relying on automatic module names

fixes #82
