+++ modbus-tcp/src/main/java/com/digitalpetri/modbus/tcp/server/NettyServerTransportConfig.java
@@ -1,7 +1,7 @@
-package com.digitalpetri.modbus.server;
+package com.digitalpetri.modbus.tcp.server;
 
 import com.digitalpetri.modbus.Modbus;
-import com.digitalpetri.modbus.Netty;
+import com.digitalpetri.modbus.tcp.Netty;
 import io.netty.channel.EventLoopGroup;
 import java.util.concurrent.ExecutorService;
 import java.util.function.Consumer;
