+++ modbus-tests/pom.xml
@@ -12,6 +12,7 @@
   <artifactId>modbus-tests</artifactId>
 
   <properties>
+    <javaModuleName>com.digitalpetri.modbus.test</javaModuleName>
     <maven.compiler.source>17</maven.compiler.source>
     <maven.compiler.target>17</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
