+++ modbus-tests/src/test/java/com/digitalpetri/modbus/test/ClientServerIT.java
@@ -1,4 +1,4 @@
-package modbus;
+package com.digitalpetri.modbus.test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
