+++ modbus/pom.xml
@@ -13,6 +13,7 @@
   <artifactId>modbus</artifactId>
 
   <properties>
+    <javaModuleName>com.digitalpetri.modbus</javaModuleName>
     <maven.compiler.source>17</maven.compiler.source>
     <maven.compiler.target>17</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
